requirements.txt
appdirs==1.4.4
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
feedparser==6.0.2
furl==2.1.0
idna==2.10
Jinja2==2.11.2
MarkupSafe==1.1.1
orderedmultidict==1.0.1
Pillow==8.0.1
psutil==5.8.0
pyee==7.0.4
pyppeteer==0.2.2
requests==2.25.1
sgmllib3k==1.0.0
six==1.15.0
soupsieve==2.1
tqdm==4.54.1
urllib3==1.26.2
websockets==8.1
